Move ConsensusAuthNRecv out to separate file
