+++ src/component.rs
@@ -16,7 +16,6 @@
 // License along with this program.  If not, see
 // <https://www.gnu.org/licenses/>.
 
-use std::collections::HashMap;
 use std::convert::Infallible;
 use std::convert::TryFrom;
 use std::fmt::Debug;
@@ -30,10 +29,8 @@ use std::net::SocketAddr;
 use std::str::from_utf8;
 use std::str::Utf8Error;
 use std::sync::Arc;
-use std::sync::RwLock;
 use std::thread::JoinHandle;
 
-use constellation_auth::authn::AuthNMsgRecv;
 use constellation_auth::authn::PassthruMsgAuthN;
 use constellation_auth::authn::SessionAuthN;
 use constellation_auth::authn::TestAuthN;
@@ -70,7 +67,6 @@ use constellation_channels::resolve::cache::ThreadedNSNameCaches;
 use constellation_channels::resolve::MixedResolver;
 use constellation_channels::unix::UnixSocketAddr;
 use constellation_common::codec::DatagramCodec;
-use constellation_common::error::MutexPoison;
 use constellation_common::net::DatagramXfrm;
 use constellation_common::net::DatagramXfrmCreate;
 use constellation_common::net::DatagramXfrmCreateParam;
@@ -84,12 +80,9 @@ use constellation_consensus_common::parties::StaticParties;
 use constellation_consensus_common::proto::ConsensusProto;
 use constellation_consensus_common::proto::ConsensusProtoRounds;
 use constellation_consensus_common::proto::SharedConsensusProto;
-use constellation_consensus_common::round::RoundMsg;
 use constellation_consensus_common::round::RoundsAdvance;
-use constellation_consensus_common::round::RoundsRecv;
 use constellation_consensus_common::round::RoundsSetParties;
 use constellation_consensus_common::state::ProtoState;
-use constellation_consensus_common::state::RoundResultReporter;
 #[cfg(feature = "standalone")]
 use constellation_pbft::msgs::PBFTMsgPERCodec;
 #[cfg(feature = "standalone")]
@@ -118,6 +111,7 @@ use crate::config::ConsensusConfig;
 use crate::config::PartiesConfig;
 #[cfg(feature = "standalone")]
 use crate::config::StandaloneConfig;
+use crate::recv::ConsensusAuthNRecv;
 use crate::state::StateThread;
 
 /// Index used to identify principals in the stream.
@@ -141,22 +135,6 @@ pub struct StringPrincipalCodec;
 // number stream.
 pub struct AscendingCount {
     curr: u128
-}
-
-struct ConsensusAuthNRecv<R, Reporter, RoundID, Prin, Oper, Msg>
-where
-    R: RoundsRecv<RoundID, PartyStreamIdx, Oper, Msg> + Send + Sync,
-    Reporter: RoundResultReporter<RoundID, Oper> + Send + Sync,
-    RoundID: Clone + Display + Ord + Send,
-    Prin: Display + Eq + Hash + Send,
-    Msg: RoundMsg<RoundID> + Send {
-    oper: PhantomData<Oper>,
-    msg: PhantomData<Msg>,
-    round_ids: PhantomData<RoundID>,
-    prins: Arc<RwLock<HashMap<Prin, PartyStreamIdx>>>,
-    reporter: Reporter,
-    notify: Notify,
-    rounds: R
 }
 
 pub type CompoundConsensusComponent<
@@ -342,152 +320,6 @@ impl From<PartyRoundIdx> for usize {
     #[inline]
     fn from(val: PartyRoundIdx) -> usize {
         val.0
-    }
-}
-
-unsafe impl<R, Reporter, RoundID, Prin, Oper, Msg> Send
-    for ConsensusAuthNRecv<R, Reporter, RoundID, Prin, Oper, Msg>
-where
-    R: RoundsRecv<RoundID, PartyStreamIdx, Oper, Msg> + Send + Sync,
-    Reporter: RoundResultReporter<RoundID, Oper> + Send + Sync,
-    RoundID: Clone + Display + Ord + Send,
-    Prin: Display + Eq + Hash + Send,
-    Msg: RoundMsg<RoundID> + Send
-{
-}
-
-unsafe impl<R, Reporter, RoundID, Prin, Oper, Msg> Sync
-    for ConsensusAuthNRecv<R, Reporter, RoundID, Prin, Oper, Msg>
-where
-    R: RoundsRecv<RoundID, PartyStreamIdx, Oper, Msg> + Send + Sync,
-    Reporter: RoundResultReporter<RoundID, Oper> + Send + Sync,
-    RoundID: Clone + Display + Ord + Send,
-    Prin: Display + Eq + Hash + Send,
-    Msg: RoundMsg<RoundID> + Send
-{
-}
-
-impl<R, Reporter, RoundID, Prin, Oper, Msg> Clone
-    for ConsensusAuthNRecv<R, Reporter, RoundID, Prin, Oper, Msg>
-where
-    R: Clone + RoundsRecv<RoundID, PartyStreamIdx, Oper, Msg> + Send + Sync,
-    Reporter: Clone + RoundResultReporter<RoundID, Oper> + Send + Sync,
-    RoundID: Clone + Display + Ord + Send,
-    Prin: Display + Eq + Hash + Send,
-    Msg: RoundMsg<RoundID> + Send
-{
-    #[inline]
-    fn clone(&self) -> Self {
-        ConsensusAuthNRecv {
-            oper: self.oper,
-            msg: self.msg,
-            round_ids: self.round_ids,
-            prins: self.prins.clone(),
-            reporter: self.reporter.clone(),
-            notify: self.notify.clone(),
-            rounds: self.rounds.clone()
-        }
-    }
-}
-
-impl<R, Reporter, RoundID, Prin, Oper, Msg>
-    ConsensusAuthNRecv<R, Reporter, RoundID, Prin, Oper, Msg>
-where
-    R: RoundsRecv<RoundID, PartyStreamIdx, Oper, Msg> + Send + Sync,
-    Reporter: RoundResultReporter<RoundID, Oper> + Send + Sync,
-    RoundID: Clone + Display + Ord + Send,
-    Prin: Display + Eq + Hash + Send,
-    Msg: RoundMsg<RoundID> + Send
-{
-    #[inline]
-    fn create(
-        reporter: Reporter,
-        rounds: R,
-        notify: Notify
-    ) -> Self {
-        ConsensusAuthNRecv {
-            oper: PhantomData,
-            msg: PhantomData,
-            round_ids: PhantomData,
-            prins: Arc::new(RwLock::new(HashMap::new())),
-            reporter: reporter,
-            notify: notify,
-            rounds: rounds
-        }
-    }
-
-    #[inline]
-    fn set_parties<I>(
-        &mut self,
-        prins: I
-    ) -> Result<(), MutexPoison>
-    where
-        I: Iterator<Item = (PartyStreamIdx, Prin)> {
-        let mut guard = self.prins.write().map_err(|_| MutexPoison)?;
-
-        *guard = prins.map(|(a, b)| (b, a)).collect();
-
-        Ok(())
-    }
-}
-
-impl<R, Reporter, RoundID, Prin, Oper, Msg> Drop
-    for ConsensusAuthNRecv<R, Reporter, RoundID, Prin, Oper, Msg>
-where
-    R: RoundsRecv<RoundID, PartyStreamIdx, Oper, Msg> + Send + Sync,
-    Reporter: RoundResultReporter<RoundID, Oper> + Send + Sync,
-    RoundID: Clone + Display + Ord + Send,
-    Prin: Display + Eq + Hash + Send,
-    Msg: RoundMsg<RoundID> + Send
-{
-    fn drop(&mut self) {
-        if let Err(err) = self.notify.notify() {
-            error!(target: "consensus-recv-authn-msg",
-                   "error notifying sender: {}",
-                   err);
-        }
-    }
-}
-
-impl<R, Reporter, RoundID, Prin, Oper, Msg> AuthNMsgRecv<Prin, Msg>
-    for ConsensusAuthNRecv<R, Reporter, RoundID, Prin, Oper, Msg>
-where
-    R: RoundsRecv<RoundID, PartyStreamIdx, Oper, Msg> + Send + Sync,
-    Reporter: RoundResultReporter<RoundID, Oper> + Send + Sync,
-    RoundID: Clone + Display + Ord + Send,
-    Prin: Display + Eq + Hash + Send,
-    Msg: RoundMsg<RoundID> + Send
-{
-    type RecvError = MutexPoison;
-
-    /// Receive an authenticated message.
-    fn recv_auth_msg(
-        &mut self,
-        prin: &Prin,
-        msg: Msg
-    ) -> Result<(), Self::RecvError> {
-        let guard = self.prins.read().map_err(|_| MutexPoison)?;
-
-        match guard.get(prin) {
-            Some(party) => {
-                if let Err(err) =
-                    self.rounds.recv(&mut self.reporter, party, msg)
-                {
-                    warn!(target: "consensus-recv-authn-msg",
-                          "error receiving message from {}: {}",
-                          prin, err)
-                }
-
-                self.notify.notify()
-            }
-            None => {
-                warn!(target: "consensus-recv-authn-msg",
-                      "discarding message from unknown principal {}",
-                      prin);
-
-                Ok(())
-            }
-        }
     }
 }
