+++ src/lib.rs
@@ -29,4 +29,5 @@
 pub mod component;
 pub mod config;
 
+mod recv;
 mod state;
